test(auth): split SSO regression coverage
